+++ db.json
@@ -14,12 +14,14 @@
         "leader1"
       ],
       "following": [
-        "leader11",
         "leader9",
-        "leader2",
         "leader3",
         "leader8",
-        "leader1"
+        "leader1",
+        "leader4",
+        "leader5",
+        "leader11",
+        "leader6"
       ],
       "accounts": [
         "acc1",
@@ -44,7 +46,6 @@
       "followers": [
         "copier1",
         "copier3",
-        "leader1",
         "leader2"
       ],
       "following": [
@@ -146,7 +147,8 @@
       "profilePicture": "https://images.unsplash.com/photo-1519085360753-af0119f7cbe7",
       "userType": "leader",
       "followers": [
-        "copier6"
+        "copier6",
+        "leader1"
       ],
       "following": [],
       "accounts": [
@@ -164,7 +166,8 @@
       "userType": "leader",
       "followers": [
         "copier7",
-        "copier8"
+        "copier8",
+        "leader1"
       ],
       "following": [],
       "accounts": [
@@ -181,7 +184,8 @@
       "profilePicture": "https://images.unsplash.com/photo-1506794778202-cad84cf45f1d",
       "userType": "leader",
       "followers": [
-        "copier9"
+        "copier9",
+        "leader1"
       ],
       "following": [],
       "accounts": [
@@ -482,8 +486,7 @@
       },
       "copiers": [
         "copier1",
-        "copier2",
-        "leader1"
+        "copier2"
       ],
       "isActive": true,
       "createdAt": "2024-01-01T00:00:00Z",
@@ -692,8 +695,7 @@
       },
       "copiers": [
         "copier4",
-        "copier5",
-        "leader1"
+        "copier5"
       ],
       "isActive": true,
       "createdAt": "2024-01-12T00:00:00Z",
@@ -735,8 +737,7 @@
       },
       "copiers": [
         "copier8",
-        "copier9",
-        "leader1"
+        "copier9"
       ],
       "isActive": true,
       "createdAt": "2024-01-14T00:00:00Z",
@@ -1041,28 +1042,6 @@
       "updatedAt": "2024-01-27T00:00:00Z"
     },
     {
-      "id": "leader1-strat11",
-      "copierId": "leader1",
-      "leaderId": "leader5",
-      "strategyId": "strat11",
-      "copierAccountId": "default",
-      "status": "active",
-      "copySize": 1000,
-      "createdAt": "2025-02-13T12:09:39.722Z",
-      "updatedAt": "2025-02-13T12:09:39.722Z"
-    },
-    {
-      "id": "leader1-strat13",
-      "copierId": "leader1",
-      "leaderId": "leader6",
-      "strategyId": "strat13",
-      "copierAccountId": "default",
-      "status": "active",
-      "copySize": 1000,
-      "createdAt": "2025-02-13T12:15:13.131Z",
-      "updatedAt": "2025-02-13T12:15:13.131Z"
-    },
-    {
       "id": "leader1-strat4",
       "copierId": "leader1",
       "leaderId": "leader2",
@@ -1074,17 +1053,6 @@
       "updatedAt": "2025-02-13T13:25:24.482Z"
     },
     {
-      "id": "leader1-strat1",
-      "copierId": "leader1",
-      "leaderId": "leader1",
-      "strategyId": "strat1",
-      "copierAccountId": "default",
-      "status": "active",
-      "copySize": 1000,
-      "createdAt": "2025-02-13T13:25:27.472Z",
-      "updatedAt": "2025-02-13T13:25:27.472Z"
-    },
-    {
       "id": "leader1-strat3",
       "copierId": "leader1",
       "leaderId": "leader1",
@@ -1092,8 +1060,8 @@
       "copierAccountId": "default",
       "status": "active",
       "copySize": 1000,
-      "createdAt": "2025-02-13T13:25:31.810Z",
-      "updatedAt": "2025-02-13T13:25:31.810Z"
+      "createdAt": "2025-02-13T14:17:59.717Z",
+      "updatedAt": "2025-02-13T14:17:59.717Z"
     }
   ],
   "posts": [
