+++ src/pages/discover/components/LeadersSection/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./LeadersSection";
