+++ src/pages/discover/components/Search/Search.css
@@ -1,0 +1,30 @@
+.discover__search {
+  position: relative;
+  margin: 1rem 0;
+}
+
+.discover__search-input {
+  width: 100%;
+  padding: 0.75rem 1rem;
+  border: 1px solid var(--border-color);
+  border-radius: 8px;
+  font-size: 1rem;
+  background-color: var(--background-color);
+  color: var(--text-color);
+}
+
+.discover__search-ai {
+  position: absolute;
+  right: 0.5rem;
+  top: 50%;
+  transform: translateY(-50%);
+  background: none;
+  border: none;
+  cursor: pointer;
+  padding: 0.25rem;
+}
+
+.discover__search-ai img {
+  width: 24px;
+  height: 24px;
+}
